跑通测试

+++ WebCrawler/src/main/java/com/blackleaf/webcrawler/dao/PageDao.java
@@ -3,5 +3,5 @@ package com.blackleaf.webcrawler.dao;
 import com.blackleaf.webcrawler.domain.Page;
 
 public interface PageDao {
-	public int insertPage(Page page);
+	public long insertPage(Page page);
 }
