+++ WebCrawler/src/main/java/com/blackleaf/webcrawler/service/PageService.java
@@ -3,5 +3,5 @@ package com.blackleaf.webcrawler.service;
 import com.blackleaf.webcrawler.domain.Page;
 
 public interface PageService {
-	public int insertPage(Page page);
+	public long insertPage(Page page);
 }
